requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
PyJWT==2.10.1
python-dotenv==1.2.1
aiosqlite==0.20.0
PyYAML>=6.0.1
requests>=2.31.0
python-multipart==0.0.9
cryptography>=43.0.0
ipfshttpclient>=0.8.0a2
